Blackfin: kgdb: make kgdb_arch_imask unsigned long to match ex_mask
git-svn-id: svn://localhost/svn/linux-kernel/trunk@9546 526b6c2d-f592-4532-a319-5dd88ccb003d

+++ arch/blackfin/kernel/kgdb.c
@@ -346,7 +346,7 @@ void kgdb_roundup_cpu(int cpu, unsigned long flags)
 #endif
 
 #ifdef CONFIG_IPIPE
-static int kgdb_arch_imask;
+static unsigned long kgdb_arch_imask;
 #endif
 
 void kgdb_post_primary_code(struct pt_regs *regs, int e_vector, int err_code)
